tests/wpt/web-platform-tests/layout-instability/content-visibility-hidden.html

<!DOCTYPE html>
<title>Layout Instability: content-visibility:hidden content</title>
<link rel="help" href="https://wicg.github.io/layout-instability/" />
<style>
  #target {
    content-visibility: hidden;
    contain-intrinsic-size: 1px;
    width: 100px;
  }
</style>
<div id=target>
  <div style="width: 100px; height: 100px"></div>
</div>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/util.js"></script>
<script>
promise_test(async () => {
  const watcher = new ScoreWatcher;

  // Wait for the initial render to complete.
  await waitForAnimationFrames(2);
  assert_equals(watcher.score, 0);
}, 'on-screen content-visibility:auto');

</script>
